refactor: atualizar dependências do MUI e reorganizar importações de estilos
