+++ package.json
@@ -104,7 +104,7 @@
     "@hookform/resolvers": "^5.2.1",
     "@mui/icons-material": "^7.3.1",
     "@mui/material": "^7.3.1",
-    "@mui/x-data-grid": "^8.10.2",
+    "@mui/x-data-grid": "^8.11.0",
     "@prisma/client": "^6.15.0",
     "dayjs": "^1.11.15",
     "js-cookie": "^3.0.5",
@@ -144,7 +144,7 @@
     "lint-staged": "^16.1.5",
     "prettier": "^3.6.2",
     "prisma": "^6.15.0",
-    "typescript": "^5.9.2",
-    "typedoc": "^0.28.11"
+    "typedoc": "^0.28.11",
+    "typescript": "^5.9.2"
   }
 }
